+++ src/utils.ts
@@ -40,20 +40,24 @@ export function makeURLSearchParams(params: Record<string, unknown> | string[][]
     return new URLSearchParams(string_params);
 }
 
+export function isDateField(k: string, v: unknown): boolean {
+    return typeof v === 'string' && (k.endsWith('_at') || k.endsWith('_date'));
+}
+
 
 /**
  * Convert to Date objects all values that look like valid date strings.
  **/
 export function coerceDates(payload: AnyPayload | any, options?: {
-    date_regexp?: RegExp,
+    is_date_field?: (k: string, v: unknown) => boolean,
     keep_invalid?: boolean,
 }): typeof payload {
     if (!payload) {
         return payload;
     }
 
-    const date_regexp = options?.date_regexp || /^(?:19[789]\d|20[0-4]\d)-\d{2}-\d{2}/,
-        keep_invalid = options?.keep_invalid ?? false;
+    const is_date_field = options?.is_date_field || isDateField,
+        keep_invalid = options?.keep_invalid ?? true;
 
     if (Array.isArray(payload)) {
         return (payload as AnyPayload[]).map(value => coerceDates(value, options));
@@ -64,8 +68,8 @@ export function coerceDates(payload: AnyPayload | any, options?: {
 
         if (Array.isArray(value)) {
             payload[field] = value.map(value => coerceDates(value, options));
-        } else if (typeof value === 'string' && date_regexp.test(value)) {
-            const date_value = new Date(value as string);
+        } else if (is_date_field(field, value)) {
+            const date_value = new Date(value);
             if (keep_invalid || !isNaN(+date_value)) {
                 payload[field] = date_value;
             }